Refactor API Models into Domain-Specific Modules
- Extracted all Pydantic models from `api.py` and organized them into dedicated modules: `wiki`, `common`, and `config`.
- Preserved existing validation rules and ensured all models maintain their functionality.
- Updated import statements throughout the codebase to reflect the new model locations.
- Created a comprehensive models package for improved maintainability and reusability.
- Confirmed that API functionality remains intact and all existing tests pass without modification.
- Completed Phase 5.1 of the API restructure, enhancing the overall architecture and organization of the codebase.
