+++ api/api.py
@@ -6,7 +6,7 @@ from fastapi.responses import JSONResponse, Response
 from typing import List, Optional, Dict, Literal
 import json
 from datetime import datetime
-from pydantic import BaseModel, Field
+
 import asyncio
 
 # Configure logging
@@ -35,113 +35,21 @@ app.add_middleware(
 def get_adalflow_default_root_path():
     return os.path.expanduser(os.path.join("~", ".adalflow"))
 
-# --- Pydantic Models ---
-class WikiPage(BaseModel):
-    """
-    Model for a wiki page.
-    """
-    id: str
-    title: str
-    content: str
-    filePaths: List[str]
-    importance: str # Should ideally be Literal['high', 'medium', 'low']
-    relatedPages: List[str]
-
-class ProcessedProjectEntry(BaseModel):
-    id: str  # Filename
-    owner: str
-    repo: str
-    name: str  # owner/repo
-    repo_type: str # Renamed from type to repo_type for clarity with existing models
-    submittedAt: int # Timestamp
-    language: str # Extracted from filename
-
-class RepoInfo(BaseModel):
-    owner: str
-    repo: str
-    type: str
-    token: Optional[str] = None
-    localPath: Optional[str] = None
-    repoUrl: Optional[str] = None
-
-
-class WikiSection(BaseModel):
-    """
-    Model for the wiki sections.
-    """
-    id: str
-    title: str
-    pages: List[str]
-    subsections: Optional[List[str]] = None
-
-
-class WikiStructureModel(BaseModel):
-    """
-    Model for the overall wiki structure.
-    """
-    id: str
-    title: str
-    description: str
-    pages: List[WikiPage]
-    sections: Optional[List[WikiSection]] = None
-    rootSections: Optional[List[str]] = None
-
-class WikiCacheData(BaseModel):
-    """
-    Model for the data to be stored in the wiki cache.
-    """
-    wiki_structure: WikiStructureModel
-    generated_pages: Dict[str, WikiPage]
-    repo_url: Optional[str] = None  #compatible for old cache
-    repo: Optional[RepoInfo] = None
-    provider: Optional[str] = None
-    model: Optional[str] = None
-
-class WikiCacheRequest(BaseModel):
-    """
-    Model for the request body when saving wiki cache.
-    """
-    repo: RepoInfo
-    language: str
-    wiki_structure: WikiStructureModel
-    generated_pages: Dict[str, WikiPage]
-    provider: str
-    model: str
-
-class WikiExportRequest(BaseModel):
-    """
-    Model for requesting a wiki export.
-    """
-    repo_url: str = Field(..., description="URL of the repository")
-    pages: List[WikiPage] = Field(..., description="List of wiki pages to export")
-    format: Literal["markdown", "json"] = Field(..., description="Export format (markdown or json)")
-
-# --- Model Configuration Models ---
-class Model(BaseModel):
-    """
-    Model for LLM model configuration
-    """
-    id: str = Field(..., description="Model identifier")
-    name: str = Field(..., description="Display name for the model")
-
-class Provider(BaseModel):
-    """
-    Model for LLM provider configuration
-    """
-    id: str = Field(..., description="Provider identifier")
-    name: str = Field(..., description="Display name for the provider")
-    models: List[Model] = Field(..., description="List of available models for this provider")
-    supportsCustomModel: Optional[bool] = Field(False, description="Whether this provider supports custom models")
-
-class ModelConfig(BaseModel):
-    """
-    Model for the entire model configuration
-    """
-    providers: List[Provider] = Field(..., description="List of available model providers")
-    defaultProvider: str = Field(..., description="ID of the default provider")
-
-class AuthorizationConfig(BaseModel):
-    code: str = Field(..., description="Authorization code")
+# Import models from the models package
+from api.models import (
+    WikiPage,
+    WikiSection,
+    WikiStructureModel,
+    WikiCacheData,
+    WikiCacheRequest,
+    WikiExportRequest,
+    ProcessedProjectEntry,
+    RepoInfo,
+    Model,
+    Provider,
+    ModelConfig,
+    AuthorizationConfig
+)
 
 from api.config import configs, WIKI_AUTH_MODE, WIKI_AUTH_CODE
