+++ api/models/__init__.py
@@ -4,3 +4,45 @@ Data models package.
 This package contains Pydantic models for API requests,
 responses, and internal data structures.
 """
+
+# Import all models for easy access
+from .wiki import (
+    WikiPage,
+    WikiSection,
+    WikiStructureModel,
+    WikiCacheData,
+    WikiCacheRequest,
+    WikiExportRequest
+)
+
+from .common import (
+    ProcessedProjectEntry,
+    RepoInfo
+)
+
+from .config import (
+    Model,
+    Provider,
+    ModelConfig,
+    AuthorizationConfig
+)
+
+__all__ = [
+    # Wiki models
+    "WikiPage",
+    "WikiSection", 
+    "WikiStructureModel",
+    "WikiCacheData",
+    "WikiCacheRequest",
+    "WikiExportRequest",
+    
+    # Common models
+    "ProcessedProjectEntry",
+    "RepoInfo",
+    
+    # Configuration models
+    "Model",
+    "Provider", 
+    "ModelConfig",
+    "AuthorizationConfig"
+]
