+++ memory-bank/progress.md
@@ -1,6 +1,6 @@
 # Progress Tracking - DeepWiki API Restructure
 
-## Overall Progress: 66.7% Complete (8/12 Phases)
+## Overall Progress: 75.0% Complete (9/12 Phases)
 
 ### Phase Status Overview
 - ✅ **Phase 1.1**: Directory Structure - **COMPLETED** (100%)
@@ -12,11 +12,30 @@
 - ✅ **Phase 3.2**: Chat Pipeline - **COMPLETED** (100%)
 - ✅ **Phase 4.1**: Chat Service - **COMPLETED** (100%)
 - ✅ **Phase 4.2**: Project Service - **COMPLETED** (100%)
-- 🔴 **Phase 5.1**: Models - **NOT STARTED** (0%)
+- ✅ **Phase 5.1**: Models - **COMPLETED** (100%)
 - 🔴 **Phase 5.2**: Endpoints - **NOT STARTED** (0%)
 - 🔴 **Phase 5.3**: App Configuration - **NOT STARTED** (0%)
 
 ## Recent Achievements
+
+### Phase 5.1: Models ✅ COMPLETED (2025-08-28)
+**Major Milestone**: Successfully extracted all Pydantic models from `api.py` and organized them into domain-specific modules.
+
+**Key Accomplishments**:
+- ✅ **Complete Extraction**: Extracted all Pydantic models from 634-line api.py file
+- ✅ **Domain Organization**: Organized models into logical domains (wiki, common, config)
+- ✅ **Validation Preservation**: Maintained 100% of existing validation rules and field definitions
+- ✅ **Import Updates**: Updated all import statements in api.py to use new model locations
+- ✅ **Package Structure**: Created comprehensive models package with proper imports
+- ✅ **Functionality Verification**: Confirmed API functionality remains completely intact
+- ✅ **Test Compatibility**: All existing tests continue to pass without modification
+- ✅ **Clean Architecture**: Established clear separation of concerns for model management
+
+**Technical Impact**:
+- **Maintainability**: Significant improvement over monolithic model definitions
+- **Organization**: Clear domain separation for better code navigation
+- **Reusability**: Models can now be imported independently from other modules
+- **Code Quality**: Professional-grade model organization and structure
 
 ### Phase 4.2: Project Service ✅ COMPLETED (2025-08-27)
 **Major Milestone**: Successfully extracted all project processing logic from `data_pipeline.py` into a dedicated service layer.
@@ -58,8 +77,8 @@
 - ✅ **4.1 Chat Service**: Complete service layer architecture with business logic extraction
 - ✅ **4.2 Project Service**: Complete project processing service with full functionality preservation
 
-### Phase 5: API Layer (0% Complete)
-- 🔴 **5.1 Models**: Pydantic models and data structures
+### Phase 5: API Layer (33.3% Complete)
+- ✅ **5.1 Models**: Pydantic models and data structures
 - 🔴 **5.2 Endpoints**: FastAPI endpoints and routing
 - 🔴 **5.3 App Configuration**: Application setup and configuration
 
@@ -79,13 +98,13 @@
 
 ## Next Steps
 
-### Immediate Priority: Phase 5.1 - Models
-**Focus**: Implement Pydantic models and data structures for the API layer
+### Immediate Priority: Phase 5.2 - Endpoints
+**Focus**: Implement FastAPI endpoints and routing for the API layer
 **Timeline**: Week 5 of restructure
 **Dependencies**: All previous phases completed ✅
 
 ### Upcoming Phases
-1. **Phase 5.2**: Endpoints implementation
+1. **Phase 5.2**: Endpoints implementation (Next Priority)
 2. **Phase 5.3**: App configuration
 3. **Phase 6**: Data layer implementation
 4. **Phase 7**: Utilities and integration
@@ -137,6 +156,6 @@
 
 **Phase 4.2 (Project Service) has been successfully completed**, marking a major milestone in the API restructure. The project service provides a comprehensive, production-ready implementation that maintains all existing functionality while significantly improving maintainability and extensibility.
 
-**Current Status**: The system now has a complete service layer architecture with both chat and project services fully implemented. All core components are extracted, unified, and working correctly. The foundation is solid for implementing the remaining phases.
+**Current Status**: The system now has a complete service layer architecture with both chat and project services fully implemented, plus a comprehensive models layer with domain-specific organization. All core components are extracted, unified, and working correctly. The foundation is solid for implementing the remaining phases.
 
-**Next Phase**: Ready to begin Phase 5.1 (Models) to implement the API layer data structures and models.
+**Next Phase**: Ready to begin Phase 5.2 (Endpoints) to implement the FastAPI endpoints and routing for the API layer.
