+++ memory-bank/task-list.md
@@ -18,7 +18,7 @@
 | TASK008 | Phase 3.2: Chat Pipeline | 🟢 **COMPLETED** | 100% | Week 3 |
 | TASK009 | Phase 4.1: Chat Service | 🟢 **COMPLETED** | 100% | Week 4 |
 | TASK010 | Phase 4.2: Project Service | 🔴 Not Started | 0% | Week 4 |
-| TASK011 | Phase 5.1: Models | 🔴 Not Started | 0% | Week 5 |
+| TASK011 | Phase 5.1: Models | ✅ Completed | 100% | Week 5 |
 | TASK012 | Phase 5.2: Endpoints | 🔴 Not Started | 0% | Week 5 |
 | TASK013 | Phase 5.3: App Configuration | 🔴 Not Started | 0% | Week 5 |
 | TASK014 | Phase 6.1: Data Processing | 🔴 Not Started | 0% | Week 6 |
