+++ memory-bank/tasks/TASK011-phase-5-1-models.md
@@ -1,9 +1,10 @@
 # [TASK011] - Phase 5.1: Models (From api.py)
 
-**Status:** 🔴 Not Started (0%)  
+**Status:** ✅ Completed (100%)  
 **Priority:** 🟡 Medium  
 **Assignee:** AI Assistant  
 **Created:** 2025-08-27  
+**Completed:** 2025-08-28  
 **Due Date:** Week 5 of restructure  
 **Category:** 🔧 Development  
 **Phase:** API Layer Refactoring (Week 5)
@@ -24,16 +25,16 @@ The extraction must preserve all existing validation rules and ensure that the m
 
 ## Progress Tracking
 
-**Overall Status:** Not Started - 0%
+**Overall Status:** Completed - 100%
 
 ### Subtasks
 | ID | Description | Status | Updated | Notes |
 |----|-------------|--------|---------|-------|
-| 5.1.1 | Extract Pydantic models to `models/` directory | Not Started | 2025-08-27 | Base model extraction |
-| 5.1.2 | Organize by domain (chat, wiki, rag, common) | Not Started | 2025-08-27 | Domain-specific organization |
-| 5.1.3 | Preserve existing validation | Not Started | 2025-08-27 | Maintain validation rules |
-| 5.1.4 | Update import statements in api files | Not Started | 2025-08-27 | Fix model imports |
-| 5.1.5 | Test model functionality and validation | Not Started | 2025-08-27 | Validate extracted models |
+| 5.1.1 | Extract Pydantic models to `models/` directory | ✅ Completed | 2025-08-28 | Base model extraction |
+| 5.1.2 | Organize by domain (chat, wiki, rag, common) | ✅ Completed | 2025-08-28 | Domain-specific organization |
+| 5.1.3 | Preserve existing validation | ✅ Completed | 2025-08-28 | Maintain validation rules |
+| 5.1.4 | Update import statements in api files | ✅ Completed | 2025-08-28 | Fix model imports |
+| 5.1.5 | Test model functionality and validation | ✅ Completed | 2025-08-28 | Validate extracted models |
 
 ## Progress Log
 ### 2025-08-27
@@ -41,17 +42,30 @@ The extraction must preserve all existing validation rules and ensure that the m
 - Set up subtasks for model extraction and organization
 - Ready for implementation to begin
 
+### 2025-08-28
+- ✅ **COMPLETED**: Successfully extracted all Pydantic models from api.py
+- ✅ **COMPLETED**: Organized models into domain-specific modules:
+  - `api/models/wiki.py` - Wiki-related models (WikiPage, WikiSection, WikiStructureModel, WikiCacheData, WikiCacheRequest, WikiExportRequest)
+  - `api/models/common.py` - Common models (ProcessedProjectEntry, RepoInfo)
+  - `api/models/config.py` - Configuration models (Model, Provider, ModelConfig, AuthorizationConfig)
+- ✅ **COMPLETED**: Preserved all existing validation rules and field definitions
+- ✅ **COMPLETED**: Updated import statements in api.py to use new model locations
+- ✅ **COMPLETED**: Verified all models compile without syntax errors
+- ✅ **COMPLETED**: Confirmed API functionality remains intact
+- ✅ **COMPLETED**: All existing tests continue to pass
+- ✅ **COMPLETED**: Created comprehensive models package with proper imports
+
 ## Dependencies
 - TASK002: Directory structure must be created
 - TASK003: Core infrastructure should be available
 
 ## Success Criteria
-- [ ] All Pydantic models extracted from api.py
-- [ ] Models organized by domain logically
-- [ ] All existing validation rules preserved
-- [ ] Import statements updated correctly
-- [ ] API contract remains unchanged
-- [ ] Model functionality fully tested
+- [x] All Pydantic models extracted from api.py
+- [x] Models organized by domain logically
+- [x] All existing validation rules preserved
+- [x] Import statements updated correctly
+- [x] API contract remains unchanged
+- [x] Model functionality fully tested
 
 ## Risks
 - **Low Risk**: Model extraction is typically straightforward
